+++ PayPalMessaging/src/main/java/com/braintreepayments/api/paypalmessaging/PayPalMessagingView.kt
@@ -1,4 +1,0 @@
-package com.braintreepayments.api.paypalmessaging
-class PayPalMessagingView {
-    // TODO: Add functionality in a future PR
-}