+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+
+* PayPalMessaging (BETA)
+  * Add `PayPalMessagingRequest`, `PayPalMessagingColor`, `PayPalMessagingLogoType`, `PayPalMessagingOfferType`, `PayPalMessagingPlacement`, `PayPalMessagingTextAlignment`, and `PayPalMessagingListener`
+  * Add `PayPalMessagingView` to display PayPal messages to promote offers such as Pay Later and PayPal Credit to customers.
+    * To get started call `PayPalMessagingView#start(Context)` with an optional `PayPalMessagingRequest`
+
 ## 4.41.0 (2024-01-18)
 
 * PayPal